Move MultilingualStringMapper to sobek-common, to make it reusable from sobek-repository
