+++ sobek-common/pom.xml
@@ -89,6 +89,11 @@
             <artifactId>spring-boot-starter-webclient</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>org.mapstruct</groupId>
+            <artifactId>mapstruct</artifactId>
+        </dependency>
+
 
     </dependencies>
 </project>
